requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-extensions==2.0.7
djangorestframework==3.8.2
idna==2.6
psycopg2==2.7.4
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
